canvas_unzip: deal with empty archives
zip or tar files with no contents are not necessarily identified
as such by `file`. if we get a generic application/octet-stream,
fall back on the extension.

this error is manifested in Canvas packages that contain an
empty web_resources folder. if the folder is not present, or
contains any files, the problem does not occur.

test plan:
 - on a Linux portal, attempt to import the package from
   the ticket

fixes ADMIN-1573

Change-Id: Ic796657f7993d06f7b5e09a845aaef714d966b8e
Reviewed-on: https://gerrit.instructure.com/176174
Tested-by: Jenkins
QA-Review: Carl Kibler <ckibler@instructure.com>
Product-Review: Jeremy Stanley <jeremy@instructure.com>
Reviewed-by: James Williams <jamesw@instructure.com>