Fix client_credentials grant tokens
Erroneously checking the jti as if it was a nonce.
This allows the tokens to be used more often than
once.

closes PLAT-4084

Test Plan:

Generate a ccg token and attempt to make a call
a few times. Note that it works. Make sure that
you have redis enabled locally.

Change-Id: I4eeed1019ac9ca04956713ed84a2a922b4ffdde0
Reviewed-on: https://gerrit.instructure.com/176586
Reviewed-by: Nathan Mills <nathanm@instructure.com>
QA-Review: Nathan Mills <nathanm@instructure.com>
Tested-by: Jenkins
Product-Review: Marc Phillips <mphillips@instructure.com>