Remove duplicate SetDefaultGradeDialog
refs: GRADE-1898

This is a prework refactor to make the changes for GRADE-1898 only
necessary in one place instead of two.

The only intended behavior change is to the existing
SetDefaultGradeDialog allow for an optional onChange param which is
reverse compatible if not used.

It appears that previously we were using a sledgehammer to remove the
dialog (`dialog.remove()`) and forgetting to invoke the jQuery UI API:

    @$dialog.dialog('close')

This meant that not only were we missing the teardown lifecycle events
that dialog wants us to use but we were also unable to add hooks to the
dialogclose lifecycle event. This also means we no longer needed the
one-off call to `$('set_default_grade').remove()` in
SetDefaultGradeDialog as was this essentially a noop in Gradebook and
New Gradebook and only used in Individual Gradebook. This focus callback
has been moved to Individual View specific code.

The open method option on dialog() was also a noop since we've hacked every
dialog to always focus the close button on render for accessibility.

The changes to the selenium specs are necessary because when the set
default grade accessibility spec failed, it was not possible to only run
that individual spec. It's a preferable practice to not abstract specs into
loops like this for this reason.

Test Plan:
 - set default grade still works in gradebook, new gradebook, and
   individual gradebook, and that the focus for each of them is correct
   upon closing or submission of the dialog

Change-Id: If08e8ce94b2aa0c9cf9735a8d18f30da63cbb8f3
Reviewed-on: https://gerrit.instructure.com/177563
Tested-by: Jenkins
Reviewed-by: Keith Garner <kgarner@instructure.com>
Reviewed-by: Jeremy Neander <jneander@instructure.com>
QA-Review: Derek Bender <djbender@instructure.com>
Product-Review: Derek Bender <djbender@instructure.com>