remove lower bound override calculations
When an override score exists on an enrollment, the actual percentage
is now returned for students. Previously, if a grading standard
existed for the course, the override score shown to students would be
the lower bound of the grade range that the override score was in.

closes GRADE-2006

Test Plan

- Enable grading standards.
- As a teacher, open the gradebook.
- Assign an override score to the student, taking care to make the
  score _not_ at the boundary of a grade (low end or high end).
- As a student, open the grades summary page and verify that the
  override score is the exact percentage given by the teacher.
- As the teacher again, enter in a letter override grade.
- As the student, verify that the override score is the lower bound
  of the grade given.

Change-Id: I7066a8d652224db4036dcaf6c637057eefb8aa73
Reviewed-on: https://gerrit.instructure.com/182046
Tested-by: Jenkins
Reviewed-by: Adrian Packel <apackel@instructure.com>
Reviewed-by: Derek Bender <djbender@instructure.com>
QA-Review: Gary Mei <gmei@instructure.com>
Product-Review: Keith Garner <kgarner@instructure.com>