Enhance performance for the security tab
This makes it so we show spinners until things have loaded.  It also
makes it so that we begin to fetch the data sooner and rely on our
cache a bit more after.  Not only should this increase speed perception,
but should also really enhance the speed and experience.

closes CORE-2324

Test Plan:
  - Enable CSP feature flag
  - Go to the security tab on an account/subaccount
  - You should initially see a spinner, then the page should show
    fully loaded once it disappears.

Change-Id: Ia0c4b556cfd6e1a364443fe8f800e8c9ac4e10c9
Reviewed-on: https://gerrit.instructure.com/182384
Tested-by: Jenkins
QA-Review: Tucker Mcknight <tmcknight@instructure.com>
Product-Review: Clay Diffrient <cdiffrient@instructure.com>
Reviewed-by: Brent Burgoyne <bburgoyne@instructure.com>