eslint: remove some eslint-ignore comments
We really should only be writing eslint-ignore in files where we KNOW
We don’t want to do what it is suggesting.

If it is suggesting something that we probably should do, but it was
written before us so we want to be lazy and not do anything about it,
we should not just eslint-ignore it. We should fix it or leave it there
for someone else to fix. 

Having an eslint-ignore in a file is a signal that someone who knew what
They are doing saw that error and said “no I want to do it anyway”

We really should not be having lots of cases where you need to even 
write `eslint-ignore`. A rule should either be 99.9% always followed or
we should get rid of it (and as a side note, I think we are there now)

A lot of these were leftovers from before we did that ^

Change-Id: I1e515d486550ad3118d6d74e9c824d28ebf6e714
Reviewed-on: https://gerrit.instructure.com/183786
Tested-by: Jenkins
Reviewed-by: Steven Burnett <sburnett@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>