spec: [ci-coverage] prevent global spec failures
This should ensure that global window errors are caught and reported,
but will allow the entire spec suite to continue. If a test is currently
in the queue, it will be failed. If no test is currently in the queue, a
failing test will be pushed onto the queue.

Change-Id: I7c0f032ac42f68b7cb359d105fe14df208c6bd3c
Reviewed-on: https://gerrit.instructure.com/187559
Tested-by: Jenkins
Reviewed-by: Keith Garner <kgarner@instructure.com>
Reviewed-by: Steven Burnett <sburnett@instructure.com>
QA-Review: Steven Burnett <sburnett@instructure.com>
Product-Review: Steven Burnett <sburnett@instructure.com>