strip out the code for wait_for_ajax from prod bundle
This code is for the wait_for_ajaximations stuff we use in selenium.
By putting this /*#__PURE__*/ comment here, it tells uglifyJS/terser
that it can remove it. So it is there when selenium runs but not there
in the actual bundles we send to real users

Test plan:
* wait_for_ajax and wait_for_ajaximations should work the same as before
  (Meaning the selenium build passes)
* run `RAILS_ENV=production  bundle exec rake canvas:compile_assets`
* you should not see it the 
  /dist/webpack-production/appBootstrap.bundle-<blah>.js 
  it creates.

Change-Id: Ia5f91b009108fc57e5406e4098964648fbd36e1b
Reviewed-on: https://gerrit.instructure.com/188320
Tested-by: Jenkins
Reviewed-by: Rob Orton <rob@instructure.com>
QA-Review: Rob Orton <rob@instructure.com>
Product-Review: Rob Orton <rob@instructure.com>