do not deadlock the data loader, dude
This bug resulted from the aptly-named "naive request dispatch" not
clearing pagination requests which had finished, thereby causing
additional requests in the queue to wait unnecessarily. In the case of
large numbers of request chunks and depaginated requests, the initial
page request could linger and be counted as active, even though it had
finished. If the queue is full of such requests, the remaining
depaginated requests will never be able to start.

fixes GRADE-2143

This is reproducible locally, which is good, because the setup required
to do a thorough stress test on this naively-throttled Gradebook data
loader is best scripted. To ensure nothing was missed in the fix for
this bug, here is what you will need:

 * 120 assignments
 * 120 students
 * everything assigned to everyone (no overrides)

The reasoning behind these counts is to force a condition where not all
chunks of students, nor all pages of submissions for each chunk of
students, can have an active request to the server at a given time. This
ensures that subsequent chunks and pages are queued and that ALL of them
are ultimately started and resolved.

test plan:
 1. Create a class with the above setup

 2. Visit New Gradebook
 3. Verify New Gradebook loads completely
 4. Verify all submission cells have the "ungraded" en dash (–)

 5. Visit Old Gradebook
 6. Verify Old Gradebook loads completely
 7. Verify all submission cells have the "ungraded" en dash (–)

Change-Id: I9bf582ac26587408a541a74d07e3304a7485eac8
Reviewed-on: https://gerrit.instructure.com/190259
Tested-by: Jenkins
Reviewed-by: Adrian Packel <apackel@instructure.com>
Reviewed-by: Keith Garner <kgarner@instructure.com>
QA-Review: James Butters <jbutters@instructure.com>
Product-Review: Keith Garner <kgarner@instructure.com>