Make the brand_config migrations have to run again
Fixes: CORE-2822

The primary purpose of this commit was to make it so that beta and
Prod are forced into running `BrandConfig.find_each(&:save_all_files!)`
again. This is necessary because we need to create the brotli encoded
versions of each of those files on the CDN.

For backstory on why that’s the case:
Recently, there was a change (g/188866) that will serve everything 
from the cdn from a root /br/ folder for browsers that support brotli.
All the rest of the CDN files got a brotli version uploaded to /br when
rake canvas:cdn:upload_to_s3 was ran as part of the deploy but because
these brand_config files are not there on the deploy box, we need
to upload them too). Otherwise people will get 404s for these files.

The way I actually accomplished this was that I ran some images through
a lossless png compressor so that their binary contents would change
and thus we’d get a new md5. So a nice side effect of this commit is
that these images will be smaller.


Test plan:
* check out master
* if you try to run rails server it should give you an error that you
  Need to run db migrations if you have not yet
* run db migrations
* now everything should work like normal


Change-Id: I806805c2ccbe75a0cf3a6cf6c664721380b802fc
Reviewed-on: https://gerrit.instructure.com/190330
Reviewed-by: Rob Orton <rob@instructure.com>
Product-Review: Rob Orton <rob@instructure.com>
QA-Review: Rob Orton <rob@instructure.com>
Tested-by: Jenkins