fix decaffeinate hack to work with babel-runtime
Fixes: CORE-2821

When decaffeinate transforms a file, if it sees something where in
CoffeeScript it would have used “this” inside the constructor function
before calling “super”, it has to use a hack to trick babel into
allowing it. For details of the “why” behind the hack see: 
https://github.com/decaffeinate/decaffeinate/blob/master/src/utils/babelConstructorWorkaroundLines.ts

Well, anyway, the hack that it uses works with stock babel but, we
are also using babel-runtime
(https://babeljs.io/docs/en/babel-plugin-transform-runtime),
and the regex it uses does not match with that. 

So, for example here is what the generated
Babel code would look like normally:


function () {
  return _assertThisInitialized(_this);
}
But with babel-runtime that becomes:


function() {
  return (0, _assertThisInitialized2.default)(_this)
}

And the regex it was using would work with the first but not the latter.


This change makes it so the regex can handle EITHER.


Test plan:
* go to /courses/x/users
* it should not throw a javascript errorr about:
  Uncaught TypeError: Cannot read property '1' of null
    at new CreateUsersView (roster.bundle-1b21617f94.js:8424)

Change-Id: I0e622b1c6ed1485da3e065e5b95abf2db1f54fa4
Reviewed-on: https://gerrit.instructure.com/190333
Tested-by: Jenkins
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
QA-Review: Clay Diffrient <cdiffrient@instructure.com>
Product-Review: Clay Diffrient <cdiffrient@instructure.com>