Fix the rest of the decaffeinate babel hacks
Fixes: CNVS-45856

I guess I missed some in my search and replace in 6cd5a1a3af0eb59

Original commit message:

When decaffeinate transforms a file, if it sees something where in
CoffeeScript it would have used “this” inside the constructor function
before calling “super”, it has to use a hack to trick babel into
allowing it. For details of the “why” behind the hack see:
https://github.com/decaffeinate/decaffeinate/blob/master/src/utils/babelConstructorWorkaroundLines.ts

Well, anyway, the hack that it uses works with stock babel but, we
are also using babel-runtime
(https://babeljs.io/docs/en/babel-plugin-transform-runtime),
and the regex it uses does not match with that.

So, for example here is what the generated
Babel code would look like normally:


function () {
  return _assertThisInitialized(_this);
}
But with babel-runtime that becomes:


function() {
  return (0, _assertThisInitialized2.default)(_this)
}

And the regex it was using would work with the first but not the latter.


This change makes it so the regex can handle EITHER.


Test plan:
* Go to a course on the beta environment
* Click on Pages from the course navigation, then click 
  "View all pages" if necessary
* Click on the "+Page" button
* it should work

Change-Id: I18165dad3d23417f2f8c9cfea52a8b53e23829b8
Reviewed-on: https://gerrit.instructure.com/190733
Tested-by: Jenkins
Reviewed-by: Jacob Burroughs <jburroughs@instructure.com>
Reviewed-by: Rob Orton <rob@instructure.com>
QA-Review: Rob Orton <rob@instructure.com>
Product-Review: Rob Orton <rob@instructure.com>