fix decaffeinate babel “no this b4 super” hack again
I recently had to fix all of these so they would work with babel-runtime
(in g/190733 and g/190333) 
And now I need to change them again because as-is they will not work
if we start passing webpack real esModules, which I do in my next commit

but this time rather than trying to make the regex even crazier, I 
realized I could do it much simpler. it was using a regular
expression to handle if uglify mangled the variable name “_this” with
with something else. but by looking at the minified code output to 
public/dist/webpack-production in a prod build, I realized that once
uglify sees an `eval(`, it doesn’t mangle any variable names that
are in scope to that `eval`. so we never have to handle that case.
meaning a simple eval(`_this = this`) will do just fine. and will
work for all cases: normal, when babel-runtime is enabled, and even when
babel-runtime is enabled and set to use esModules.


    Test plan:
    * Go to a course in a RAILS_ENV=production (minified) environment
    * Click on Pages from the course navigation, then click
      "View all pages" if necessary
    * Click on the "+Page" button
    * it should work

Change-Id: I980072a73a98ed808326afedf7849f1b3f6b8e09
Reviewed-on: https://gerrit.instructure.com/191411
Tested-by: Jenkins
Reviewed-by: Steven Burnett <sburnett@instructure.com>
QA-Review: Steven Burnett <sburnett@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>