Restore LTI submitting from "More" tab
Restore the functionality of the LTI app links in the "More" tab, and
ensure that submitting a file using said links succeeds.

This is a partial revert of 5fe7725d6881b6a64ea799cb7e078c16c65308e0,
which removed some code in submit_assignment.js that was necessary for
links in the "More" tab to function properly; it restores that code and
fixes it up to remove linting errors.

fixes GRADE-1679

Test plan:

Note: at the time of this writing, there are some not-inconsiderable
differences in behavior depending on whether the user submits a file
from the LTI app's specific tab or they use the app's entry in the
"More" tab.  Although the file should successfully be uploaded in both
cases, submission comments the user has typed appear to be lost when
submitting from the app-specific tab.  patchset makes no attempt to iron
out those differences; it only seeks to ensure apps can submit when
accessed from the "More" tab.

- Set up 4 or more LTI assignments for an account/course
  - They can all be copies of a single app, e.g. Dropbox:
  - https://www.eduappcenter.com/configurations/slv9zn9yk7ik5jpp.xml
    is the configuration URL
- Create an assignment allowing file uploads
- As a student, open the assignment and click "Submit Assignment"
  - Click the "More" tab
  - Links for all your LTI apps should appear there
    - When clicked, they should open a frame allowing you to select
      a file as appropriate for the selected app
    - Select a file and click "Submit Assignment"
    - The submission process should complete without any hitches
  - Make sure that submitting via the app-specific tabs continues to
    save the file properly (but see the note above)

Change-Id: I92e8a60dfa86ceaf7373ecfa6f5d3fa5328a8795
Reviewed-on: https://gerrit.instructure.com/192458
Tested-by: Jenkins
Reviewed-by: Jeremy Neander <jneander@instructure.com>
Reviewed-by: Keith Garner <kgarner@instructure.com>
QA-Review: Derek Bender <djbender@instructure.com>
Product-Review: Keith Garner <kgarner@instructure.com>