Query submission and submission histories separately
We cannot paginate off of submission histories for the most current
submission anymore, because that is now a much more expensive call that
we don't want to inflict unless we are actually looking at previous
submissions for the assignment. This will now query the current
submission for this assignment from the `SubmissionConnection` GQL api,
and only load the histories on demand from the
`SubmissionHistoriesConnection` GQL api if the previous button is
clicked.

Fixes COMMS-2092

Test Plan:
  * Jenkins passes
  * Everything on the A2 student view continues to work the same as it
    does on master

Change-Id: I8775570e56e4145a12071a8f6b438c61f3640c41
Reviewed-on: https://gerrit.instructure.com/196204
Tested-by: Jenkins
Reviewed-by: Michelle Simmons <misimmons@instructure.com>
Reviewed-by: Ryan Norton <rnorton@instructure.com>
QA-Review: Matthew Lemon <mlemon@instructure.com>
Product-Review: Landon Gilbert-Bland <lbland@instructure.com>