connect front end & back end for 'Prev' and 'Next' buttons
fixes COMMS-2080

The UI and the functionality of the 'Prev' and 'Next' buttons
already exists in Assignments 2, but they need to be connected in
order to have the buttons be fully operational in the pizza
tracker. This commit connects the front end and the back end of
the 'Prev' and 'Next' buttons so that when clicked, they cause the
assignment to cycle through all the attempts chronologically. The
buttons 'Prev,' 'Next,' and 'New Attempt' appear only when
appropriate; conditions for button appearance are detailed in the
third step of the Test Plan below.

Test Plan:
* As a teacher, create an assignment
* As a student, create many attempts on the assignment; leave them
  in various states (e.g. uploaded, submitted, graded...)
* As the student, use the pizza tracker to navigate through the
  attempts. Notice that on the first attempt, the 'Previous' button
  doesn't appear, and on the last attempt, the 'Next' button
  appears. Notice also that the 'New Attempt' button only appears
  on the last attempt.

Change-Id: I21881e245f0d0c990b5715744b7a353974d4ebc3
Reviewed-on: https://gerrit.instructure.com/197244
Tested-by: Jenkins
Reviewed-by: Steven Burnett <sburnett@instructure.com>
QA-Review: Matthew Lemon <mlemon@instructure.com>
Product-Review: Michelle Simmons <misimmons@instructure.com>