Fix bug when trying to edit a conference
Fixes: CORE-3035

This code was written poorly. It had a typo where it meant to use a 
fat arrow but used a skinny arrow so on this line it thought it was
doing “this.$” where it thought “this” was the backbone view instance
but because they used a skinny arrow this.$ was actually globalThis.$,
which happened to work pre “strict mode” because this in a global 
context would be window, and there happens to be a global variable “$”
available at window.$. but it breaks in strict mode, which is inferred
whenever a file is being interpreted as an ES module.

In strict mode if you do this.whatever in a global context, “this” will
be undefined.

Test plan:
* Create a conference.
* Attempt to edit the conference and notice that it doesn’t throw an 
  error.

Change-Id: Ia04500c17d517718272d39984793bac7bc210567
Reviewed-on: https://gerrit.instructure.com/196487
Tested-by: Jenkins
Reviewed-by: Steven Burnett <sburnett@instructure.com>
QA-Review: Steven Burnett <sburnett@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>