remove the grade and time of submission from pizza tracker
fixes COMMS-2103

In the 'Submitted' step container, the pizza tracker displays the
time and date of submission. In the 'Graded' step containter, the
pizza tracker displays the grade received. This info is being
moved to the attempt grading area in a separate ticket; this
ticket sees the removal of the date/time from the 'Submitted' step
container and of the grade info from the 'Graded step container.

Test Plan:
* As a teacher, create an assignment
* As a student, submit two attempts to the assignment
* As the teacher, grade on of the student's two attempts; leave
  the other attempt ungraded.
* As a student, note that for the ungraded attempt, the date and
  time of your submission no longer appear; the step is labeled
  only "Submitted." Note that for the graded attempt, the grade
  received no longer appears in the pizza tracker; the step is
  labeled only "Graded."

Change-Id: Ibb3e152e50054a036f4e2a4ee0d7978388d96716
Reviewed-on: https://gerrit.instructure.com/196667
Reviewed-by: Ryan Norton <rnorton@instructure.com>
Tested-by: Jenkins
Product-Review: Michelle Simmons <misimmons@instructure.com>
QA-Review: Landon Gilbert-Bland <lbland@instructure.com>