Don’t try to use sessionStorage if XHR has already returned
On the dashcard dashboard, on first render we try to render with data
cached in sessionStorage, then once the XHR to get the freshest
data comes back, we render again with the new data.

But the thing is, now that we prefetch that XHR from our rails erb,
by the time we get to this line of code, the XHR might already be
Loaded. So if that’s the case, there is no need to try to render from
sessionStorage with cached data, we can just render with the new data.


Test plan:
* the dashcard dashboard should look the same as before but it should
  Load a little faster.

Change-Id: I1132e698ddd667623bb9547150afc8042312ab02
Reviewed-on: https://gerrit.instructure.com/196666
Tested-by: Jenkins
Reviewed-by: Mysti Lilla <mysti@instructure.com>
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>