change permission for can_edit_users
the permissions for managing a user through the ui reads from
this permissions env, and needed to be changed so it matches
the actual values sent from the frontend. there was a case where
a subaccount admin could manage_students, but couldnt manage
the details of people for that account.

Test-Plan:
- ensure that a sub account admin does not see the edit icon
  in the people page
- ensure that all other admins that should be able to see the
  edit icon can

fixes: ADMIN-2731
Change-Id: I887d070310ff17d09931d990339c4a812541157e
Reviewed-on: https://gerrit.instructure.com/199160
Tested-by: Jenkins
Product-Review: Rex Fleischer <rfleischer@instructure.com>
QA-Review: Carl Kibler <ckibler@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>