don’t send xhr to folders/root on modules page
Before this commit:
If you go to the /courses/x/modules page, it will send an XHR request
to /api/v1/courses/1/folders/root when the page loads. That is needless
and wasteful. 

It’s because FilesEnv, which is required by RestrictedDialogForm, which 
is required by PublishCloud, and is only needed once you click on the 
“publish/unpublish” icon for a file requests it.

So to prevent the needless XHR request, instead of loading it always at
pageload, we just load it once you actually click on that 
“publish/unpublish” icon.

Test plan:
before checking this out:
* go to courses/x/modules
* notice that there is an XHR request to /api/v1/courses/x/folders/root
  In the network panel

after checking this out
* go to /courses/x/modules
* notice there is NOT an xhr request to /api/v1/courses/1/folders/root
* add a module item that is a file
* click on the green circle with a checkmark in it for that file to 
  Open the “Editing permissions for: <filename>” dialog
* it should work the same as before
* you should now see that there was a XHR request to 
  /api/v1/courses/x/folders/root in the network panel

(see spec coverage in context_modules_teacher_spec.rb)

Change-Id: I898b68ec50f2b1af6653be58d4265e594e26b28c
Reviewed-on: https://gerrit.instructure.com/199447
Tested-by: Jenkins
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>