spec: make a spec work when let isn’t converted to var
We are going to drop ie11. Which means that `let` and `const` won’t be
Converted to `var` anymore.
This spec had some bad code that was depending on that.

This change makes it work exactly as it was, in fact, working before

Change-Id: Ic8e88d622b5f944a504996b8fdd9028110147221
Reviewed-on: https://gerrit.instructure.com/199583
Tested-by: Jenkins
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>