allow sending messages to group members from other sections
fixes COMMS-1308

When students are restricted to viewing only members of their own
section, and a group contains students from different sections,
students cannot send a message in Conversations to their group
members in a different section when the class is selected as the
context rather than the group itself. This commit allows students
to send a message to a group member in a different section when a
class is selected as a context.

Test Plan:
* Create a course with two sections; select the option that
  prevents students from communicating with students in a section
  other than their own.
* Create a group that contains students from both sections.
* Masquerade as a student that belongs to the group; compose a
  message in Conversations. In the "course" form, choose the course
  that you just created. In the "To" form, click on the address
  book icon; select Groups; select the group that the student
  belongs to; and select a student in the group that belongs to a
  different section than the student you're masquerading as
* Fill in subject & content; send the message
* The message should send with no errors

Change-Id: I0c0ab55d37b2c01d9dfa3a71d1f8940e565df600
Reviewed-on: https://gerrit.instructure.com/200187
Tested-by: Jenkins
Reviewed-by: Ryan Norton <rnorton@instructure.com>
QA-Review: Ryan Norton <rnorton@instructure.com>
Product-Review: Michelle Simmons <misimmons@instructure.com>