enable discussion attachments by default
support has reported a lot of extra load surrounding the fact
that teachers expect students to be able to attach files to
dicussion posts by default and that the user flow to enable this
option is not immediately apparent.

we should update this to allow students to be able to attach
files to discussion posts by default.

Test Plan:
* as a teacher, create an assignment
* as a student in this assignment, navigate to the discussion page
* when creating a reply, in the bottom left of the reply panel
  should be an "Attach" button
* you should be able to attach and submit the reply

* as a teacher, navigate to the discussion page of the assignment
* in the top right, next to the new discussion button, click on
  the gear icon
* the "Attach files to discussion" option should be checked
* uncheck the "Attach files to discussion" option

* as a student, navigate to the discussions page
* when creating a reply, you should no longer have the "Attach"
  button in the bottom left and can no longer submit an attachment

fixes COMMS-2226

Change-Id: Id2df8d6cb9b9a5bc37316a7d26249dd0e663c8cc
Reviewed-on: https://gerrit.instructure.com/201823
Tested-by: Jenkins
Reviewed-by: Steven Burnett <sburnett@instructure.com>
QA-Review: Steven Burnett <sburnett@instructure.com>
Product-Review: Ryan Norton <rnorton@instructure.com>