get js_bundle ready for streaming
With this change, anything that was js_bundle’ed by the time we render
_head.html.erb will be included in the <head> so that if we are
streaming, it will get flushed to the browser earlier and the browser
knows about it as soon as possible.  But then if someone js_bundle’s
something inside the erb view, that is fine too and it will go at
the bottom of the page in foot.html.erb exactly as it was before.

Test plan:
* if you view-source, you should see that most of the
  <link rel=“preload” tags for our webpack chunks are in the <head>
  and not in the footer where they used to be
* everything should function the same, aka builds should pass

Change-Id: I6ae57546adb3b7358bc346943a1490a574c89f68
Reviewed-on: https://gerrit.instructure.com/202372
Tested-by: Jenkins
Reviewed-by: Steven Burnett <sburnett@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>